client configuration refactorings - WIP
