+++ lib/openstax/exchange/client/client_instance_base.rb
@@ -1,0 +1,17 @@
+module OpenStax
+  module Exchange
+    module Client
+
+      class ClientInstanceBase
+        def is_real?
+          raise NotImplementedError
+        end
+
+        def token
+          raise NotImplementedError
+        end
+      end
+
+    end
+  end
+end
