+++ lib/openstax/exchange/client/fake_client_configuration.rb
@@ -1,0 +1,11 @@
+module OpenStax
+  module Exchange
+    module Client
+
+      class FakeClientConfiguration
+        attr_accessor :registered_platforms
+      end
+
+    end
+  end
+end
